Added some parsing

+++ include/util.h
@@ -3,7 +3,7 @@
 
 #include <unistd.h>
 #include <stdlib.h>
-
+#include <stdio.h>
 /**
  * Reads one line from a file descriptor.
  *
@@ -12,7 +12,7 @@
  * @param max_len  Max number of bytes to read.
  * @return Number of bytes read (0 on EOF).
  */
-int get_line(int fd, char* buf, size_t maxlen);
+int get_line(int fd, char* buf, int maxlen);
 
 
 #endif
